Ana() rewritten + a basic test. More tests needed here.
