+++ Composite/.vscode/settings.json
@@ -25,5 +25,6 @@
         "typescript",
         "typescriptreact",
         "yml"
-    ]
+    ],
+    "editor.renderWhitespace": "all"
 }
