+++ Composite/Composite.Cs.Tests/AnaTests.cs
@@ -1,0 +1,50 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+
+using Xunit;
+
+using Composite;
+
+namespace Composite.Cs.Tests
+{
+    using SimpleComposite = Composite<Simple>.Composite;
+
+    public class AnaTests
+    {
+        [Fact]
+        public void UnfoldTest()
+        {
+            var scn = new Func<Simple, Simple[]>[] {
+                    x => x.Number == 1
+                        ? new [] { new Simple { Number = 2, }, new Simple { Number = 3, }}
+                        : new[] { x },
+                    x => x.Number == 2
+                        ? new [] { new Simple { Number = 4, }, new Simple { Number = 5, }}
+                        : new[] { x },
+                };
+
+            var obj = C.Composite(new[] {
+                C.Value( new Simple { Number = 1, } ),
+                C.Value( new Simple { Number = 6, } ),
+            });
+
+            var result = obj.Ana(scn);
+
+            Assert.True(result.IsComposite);
+
+            if (result is SimpleComposite compositeResult)
+            {
+                var resultArray = compositeResult.Item.ToArray();
+
+                Assert.True(resultArray.Length == 2);
+                Assert.True(resultArray[0].IsComposite);
+                Assert.True(resultArray[1].IsComposite);
+            }
+            else
+            {
+                Assert.True(false);
+            }
+        }
+    }
+}
