+++ Composite/Composite.Cs.Tests/Tests.cs
@@ -6,8 +6,6 @@ using Xunit;
 
 using Composite;
 
-using SimpleComposite = Composite.DataTypes.Composite<Composite.Cs.Tests.Simple>.Composite;
-
 namespace Composite.Cs.Tests
 {
     public class Tests
@@ -108,41 +106,6 @@ namespace Composite.Cs.Tests
             {
                 C.ToFlat(obj).Take(5).ToArray();
             });
-        }
-
-        [Fact]
-        public void AnaTest()
-        {
-            var scn = new Func<Simple, Simple[]>[] {
-                    x => x.Number == 1
-                        ? new [] { new Simple { Number = 2, }, new Simple { Number = 3, }}
-                        : new[] { x },
-                    x => x.Number == 2
-                        ? new [] { new Simple { Number = 4, }, new Simple { Number = 5, }}
-                        : new[] { x },
-                };
-
-            var obj = C.Composite(new[] {
-                C.Value( new Simple { Number = 1, } ),
-                C.Value( new Simple { Number = 6, } ),
-            });
-
-            var result = C.Ana(scn, obj);
-
-            Assert.True(result.IsComposite);
-
-            if (result is SimpleComposite compositeResult)
-            {
-                var resultArray = compositeResult.Item.ToArray();
-
-                Assert.True(resultArray.Length == 2);
-                Assert.True(resultArray[0].IsComposite);
-                Assert.True(resultArray[1].IsValue);
-            }
-            else
-            {
-                Assert.True(false);
-            }
         }
     }
 }
