+++ Composite/Composite/DataTypes.fs
@@ -1,7 +1,5 @@
 namespace Composite
 
-module DataTypes =
-
-    type 'a Composite =
-        | Value of 'a
-        | Composite of seq<Composite<'a>>
+type 'a Composite =
+    | Value of 'a
+    | Composite of seq<Composite<'a>>
